+++ src/Utility/Screenshots/FixtureManager.php
@@ -2,6 +2,7 @@
 
 namespace App\Utility\Screenshots;
 
+use App\Doctrine\ORM\Filter\TrainingAreaPeriodFilter;
 use App\Entity\AreaPeriod;
 use App\Entity\Interviewer;
 use App\Entity\User;
@@ -36,10 +37,12 @@ class FixtureManager
         $areaPeriod = $areaRepository->findOneBy(['area' => self::AREA_CODE]);
 
         if (!$areaPeriod) {
+            $now = new \DateTime();
             $areaPeriod = (new AreaPeriod())
                 ->setArea(self::AREA_CODE)
                 ->addInterviewer($interviewer)
-                ->populateMonthAndYearFromArea();
+                ->setYear($now->format('Y'))
+                ->setMonth($now->format('m'));
 
             $this->entityManager->persist($areaPeriod);
         }
@@ -80,41 +83,54 @@ class FixtureManager
         $this->entityManager->flush();
 
         // Now delete everything...
+        foreach ($interviewer->getTrainingAreaPeriods()->toArray() as $area) {
+            $this->removeArea($x);
+        }
         $this->entityManager->remove($interviewer);
         $this->entityManager->remove($user);
 
+        $this->entityManager->flush();
+
         foreach($areas as $area) {
+            $this->removeArea($area);
+        }
+
+        if ($area->getInterviewers()->isEmpty()) {
             $this->entityManager->remove($area);
+        }
 
-            foreach($area->getOtpUsers() as $otpUser) {
-                $this->entityManager->remove($otpUser);
-            }
+        $this->entityManager->flush();
+    }
 
-            foreach($area->getHouseholds() as $household) {
-                $this->entityManager->remove($household);
-                $household->getVehicles()->forAll(fn($_, Vehicle $vehicle) => $this->entityManager->remove($vehicle));
+    protected function removeArea(AreaPeriod $area): void
+    {
+        $this->entityManager->remove($area);
 
-                foreach($household->getDiaryKeepers() as $diaryKeeper) {
-                    $this->entityManager->remove($diaryKeeper);
-                    $this->entityManager->remove($diaryKeeper->getUser());
+        foreach($area->getOtpUsers() as $otpUser) {
+            $this->entityManager->remove($otpUser);
+        }
 
-                    foreach($diaryKeeper->getDiaryDays() as $day) {
-                        $this->entityManager->remove($day);
+        foreach($area->getHouseholds() as $household) {
+            $this->entityManager->remove($household);
+            $household->getVehicles()->forAll(fn($_, Vehicle $vehicle) => $this->entityManager->remove($vehicle));
 
-                        foreach($day->getJourneys() as $journey) {
-                            $this->entityManager->remove($journey);
+            foreach($household->getDiaryKeepers() as $diaryKeeper) {
+                $this->entityManager->remove($diaryKeeper);
+                $this->entityManager->remove($diaryKeeper->getUser());
 
-                            foreach($journey->getStages() as $stage) {
-                                $this->entityManager->remove($stage);
-                            }
+                foreach($diaryKeeper->getDiaryDays() as $day) {
+                    $this->entityManager->remove($day);
+
+                    foreach($day->getJourneys() as $journey) {
+                        $this->entityManager->remove($journey);
+
+                        foreach($journey->getStages() as $stage) {
+                            $this->entityManager->remove($stage);
                         }
                     }
                 }
             }
         }
-
-        $this->entityManager->flush();
-
-        // TODO: Delete area if (and only if) no interviewers remaining
     }
+
 }
